requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0

# Database
pymongo==4.6.1
dnspython==2.4.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web app
streamlit==1.29.0

# Utilities
joblib==1.3.2
tqdm==4.66.1
